+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 > The version number of fuji follows `semver` now: https://semver.org/
 
-- fix: (chat.display module) The sender player is wrong when sharing `inv` or `ender`.
-- 
+- (back module) feature: persist the `saved-location` for `/back` command.